refactor: update to use double for radius and distance calculations

+++ src/colliderNode/colliderGeometry.cpp
@@ -6,6 +6,7 @@
 #include <maya/MDrawContext.h>
 
 #include <algorithm>
+#include <cmath>
 
 namespace colliderGeometry
 {
@@ -14,7 +15,7 @@ namespace colliderGeometry
         depthPriority(MHWRender::MRenderItem::sDormantWireDepthPriority)
     {}
 
-    float getDistancePlugAsCentimeters(const MObject& node, const MObject& attribute, float defaultValue)
+    double getDistancePlugAsCentimeters(const MObject& node, const MObject& attribute, double defaultValue)
     {
         MPlug plug(node, attribute);
         if (plug.isNull())
@@ -29,12 +30,12 @@ namespace colliderGeometry
             return defaultValue;
         }
 
-        return static_cast<float>(value.asCentimeters());
+        return value.asCentimeters();
     }
 
-    MBoundingBox makeSphereBoundingBox(float radius)
+    MBoundingBox makeSphereBoundingBox(double radius)
     {
-        radius = std::max(0.0f, radius);
+        radius = std::max(0.0, radius);
 
         MPoint corner1(-radius, -radius, -radius);
         MPoint corner2(radius, radius, radius);
@@ -42,34 +43,34 @@ namespace colliderGeometry
         return MBoundingBox(corner1, corner2);
     }
 
-    void appendWireSphere(MPointArray& lineList, float radius, int segments)
+    void appendWireSphere(MPointArray& lineList, double radius, int segments)
     {
-        radius = std::max(0.0f, radius);
+        radius = std::max(0.0, radius);
         segments = std::max(4, segments);
 
-        const float pi = static_cast<float>(mathUtils::kPi);
+        const double pi = mathUtils::kPi;
 
         for (int i = 0; i < segments; ++i)
         {
-            const float theta1 = pi * static_cast<float>(i) / static_cast<float>(segments);
-            const float theta2 = pi * static_cast<float>(i + 1) / static_cast<float>(segments);
+            const double theta1 = pi * static_cast<double>(i) / static_cast<double>(segments);
+            const double theta2 = pi * static_cast<double>(i + 1) / static_cast<double>(segments);
 
             for (int j = 0; j < segments; ++j)
             {
-                const float phi1 = 2.0f * pi * static_cast<float>(j) / static_cast<float>(segments);
-                const float phi2 = 2.0f * pi * static_cast<float>(j + 1) / static_cast<float>(segments);
+                const double phi1 = 2.0 * pi * static_cast<double>(j) / static_cast<double>(segments);
+                const double phi2 = 2.0 * pi * static_cast<double>(j + 1) / static_cast<double>(segments);
 
-                const float x1 = radius * std::sin(theta1) * std::cos(phi1);
-                const float y1 = radius * std::cos(theta1);
-                const float z1 = radius * std::sin(theta1) * std::sin(phi1);
+                const double x1 = radius * std::sin(theta1) * std::cos(phi1);
+                const double y1 = radius * std::cos(theta1);
+                const double z1 = radius * std::sin(theta1) * std::sin(phi1);
 
-                const float x2 = radius * std::sin(theta1) * std::cos(phi2);
-                const float y2 = radius * std::cos(theta1);
-                const float z2 = radius * std::sin(theta1) * std::sin(phi2);
+                const double x2 = radius * std::sin(theta1) * std::cos(phi2);
+                const double y2 = radius * std::cos(theta1);
+                const double z2 = radius * std::sin(theta1) * std::sin(phi2);
 
-                const float x3 = radius * std::sin(theta2) * std::cos(phi1);
-                const float y3 = radius * std::cos(theta2);
-                const float z3 = radius * std::sin(theta2) * std::sin(phi1);
+                const double x3 = radius * std::sin(theta2) * std::cos(phi1);
+                const double y3 = radius * std::cos(theta2);
+                const double z3 = radius * std::sin(theta2) * std::sin(phi1);
 
                 lineList.append(MPoint(x1, y1, z1));
                 lineList.append(MPoint(x2, y2, z2));
